# Fit-Interview-AI/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for CV Fit & Interview AI
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ExperienceEntry {
  company: string;
  role: string;
  bullet_points: string[];
}

export interface TailoredCV {
  name: string;
  summary: string;
  experience: ExperienceEntry[];
  education: string;
  skills: string[];
}

export interface CVAnalysisResult {
  /**
   * @minimum 0
   * @maximum 100
   */
  match_score: number;
  missing_skills: string[];
  tailored_cv: TailoredCV;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface InterviewChatRequest {
  jd_text: string;
  cv_text: string;
  chat_history: ChatMessage[];
}

export interface InterviewChatResponse {
  message: string;
}

export type AnalyzeCVBody = {
  /** Extracted text content from the CV */
  cv_text: string;
  /** Job description text */
  jd_text: string;
};