# Fit-Interview-AI/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for CV Fit & Interview AI
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Parses uploaded CV PDF and compares it against the job description using AI
 * @summary Upload CV and match against JD
 */
export const AnalyzeCVBody = zod.object({
  cv_text: zod.string().describe("Extracted text content from the CV"),
  jd_text: zod.string().describe("Job description text"),
});

export const analyzeCVResponseMatchScoreMin = 0;
export const analyzeCVResponseMatchScoreMax = 100;

export const AnalyzeCVResponse = zod.object({
  match_score: zod
    .number()
    .min(analyzeCVResponseMatchScoreMin)
    .max(analyzeCVResponseMatchScoreMax),
  missing_skills: zod.array(zod.string()),
  tailored_cv: zod.object({
    name: zod.string(),
    summary: zod.string(),
    experience: zod.array(
      zod.object({
        company: zod.string(),
        role: zod.string(),
        bullet_points: zod.array(zod.string()),
      }),
    ),
    education: zod.string(),
    skills: zod.array(zod.string()),
  }),
});

/**
 * Continues the mock interview conversation with the AI interviewer
 * @summary Send a message to the mock interviewer
 */
export const InterviewChatBody = zod.object({
  jd_text: zod.string(),
  cv_text: zod.string(),
  chat_history: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant", "system"]),
      content: zod.string(),
    }),
  ),
});

export const InterviewChatResponse = zod.object({
  message: zod.string(),
});